(Other) [UI Theme] Change the background color of all UIs with the same background color as the main menu. (Other) [Code] Move the Mod Manager code from Tools to ModTools. (Other) [Code] Move Modrinth stuff to it's own folder. (Other) [Android Manifest] Add "android:usesCleartextTraffic="true" " to Android Manifest to get lwjgl3ify downloading.
