+++ app_pojavlauncher/src/main/AndroidManifest.xml
@@ -31,7 +31,8 @@
         android:process=":launcher"
         android:resizeableActivity="true"
         android:roundIcon="@mipmap/ic_launcher_round"
-        android:theme="@style/AppTheme">
+        android:theme="@style/AppTheme"
+        android:usesCleartextTraffic="true">
         <activity
             android:name=".MissingStorageActivity"
             android:exported="false" />
