+++ app_pojavlauncher/src/main/assets/components/lwjgl3/version
@@ -1,1 +1,1 @@
-1680030598790
+1680146308449
