+++ app_pojavlauncher/src/main/java/net/kdt/pojavlaunch/customcontrols/ControlDrawerData.java
@@ -1,16 +1,14 @@
 package net.kdt.pojavlaunch.customcontrols;
 
-import net.kdt.pojavlaunch.Tools;
-
-import java.util.ArrayList;
-
 import static net.kdt.pojavlaunch.customcontrols.ControlDrawerData.Orientation.DOWN;
+import static net.kdt.pojavlaunch.customcontrols.ControlDrawerData.Orientation.FREE;
 import static net.kdt.pojavlaunch.customcontrols.ControlDrawerData.Orientation.LEFT;
 import static net.kdt.pojavlaunch.customcontrols.ControlDrawerData.Orientation.RIGHT;
 import static net.kdt.pojavlaunch.customcontrols.ControlDrawerData.Orientation.UP;
-import static net.kdt.pojavlaunch.customcontrols.ControlDrawerData.Orientation.FREE;
 
-import androidx.annotation.Keep;
+import net.kdt.pojavlaunch.Tools;
+
+import java.util.ArrayList;
 
 @androidx.annotation.Keep
 public class ControlDrawerData {
