+++ app_pojavlauncher/src/main/java/net/kdt/pojavlaunch/mod/modrinth/ModrinthSearchCategories.java
@@ -1,4 +1,4 @@
-package net.kdt.pojavlaunch.mod;
+package net.kdt.pojavlaunch.mod.modrinth;
 
 public enum ModrinthSearchCategories {
     forge,
